refactor: move styled modal stuff to styled file
